turned UUID to long

+++ src/main/groovy/com/app/lot_2/domain/check/controller/dto/PositionDto.java
@@ -4,11 +4,9 @@ package com.app.lot_2.domain.check.controller.dto;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class PositionDto {
 
     @Positive
-    private BigDecimal sum;
+    private long sum;
 }
